+++ CHANGES.rst
@@ -13,6 +13,9 @@ Changelog
 2.3.2 (2019-12-05)
 ------------------
 
+- Added field to set a CSS class and placeholders for fields translations
+  [daniele]
+
 - Added Alt attribute to summary view, needed for accessibility
   [rnunez]