+++ CHANGES.rst
@@ -12,6 +12,8 @@ Changelog
   [agitator]
 - Upgrade tests to github actions
   [djay]
+- CI with Github actions: Python 2 / Python 3 and os version
+  [ksuess]
 
 2.3.2 (2019-12-05)
 ------------------